Struct lockfreehashmap::atomic::NotNull

pub struct NotNull<'t, T: 't>(Shared<'t, T>);

A wrapper around [Shared], for values that we know are not null and are safe to dereference.

Methods

impl<'t, T> NotNull<'t, T>

pub fn as_maybe_null(&self) -> MaybeNull<'t, T>

This is effectively just weakening the guarantees around this type.

pub fn as_shared(&self) -> Shared<'t, T>

pub fn deref<'f>(&'f self) -> &'t T

Stronger version of Deref. This returns &'t T, rather than &'f T.

pub unsafe fn drop(self)

Drop the underlying value behind this pointer.

Unsafe

This is unsafe because only one thread can call this for any single underlying value.