Enum lockfreehashmap::map_inner::ValueSlot

pub enum ValueSlot<'v, V: 'v> {
    Value(V),
    Tombstone,
    ValuePrime(&'v ValueSlot<'v, V>),
    SeeNewTable,
}

The hash map is implemented as an array of key-value pairs, where each key and value can be one of several states. This enum represents the various states that a value can be in, excluding the null/empty state.

Variants
Value(V)

A value has been inserted into the table.

Tombstone

This state represents that a key has been inserted but then removed.

ValuePrime(&'v ValueSlot<'v, V>)

The table is being resized currently and the value here still needs to be inserted into the newer table.

SeeNewTable

This state represents one of two things: 1) This was a ValueSlot::Tombstone(_) slot that is now taken. There is a newer (resized) table that should be used if this value slot was needed. 2) This used to be a ValueSlot::Value(_) slot that has now been copied into the newer table. This is the final state for any ValueSlot.

Methods

impl<'v, V> ValueSlot<'v, V>

pub fn is_tombstone(&self) -> bool

Returns true if and only if the ValueSlot has discriminant Tombstone.

pub fn is_valueprime(&self) -> bool

Returns true if and only if the ValueSlot has discriminant ValuePrime.

pub fn is_value(&self) -> bool

Returns true if and only if the ValueSlot has discriminant Value.

pub fn is_tombprime(&self) -> bool

Returns true if and only if the ValueSlot has discriminant SeeNewTable.

pub fn is_prime(&self) -> bool

Returns true if and only if the ValueSlot has either discriminant ValuePrime or SeeNewTable.

pub fn as_inner(value: Option<&Self>) -> Option<&V>

Return an Option reference to the inner value of generic type V.