Crate lockfreehashmap

LockFreeHashMap

This is an implementation of the lock-free hash map created by Dr. Cliff Click.

Originally, this implementation here and recently here was created for Java, using garbage collection where necessary. This library is a Rust implementation, using epoch-based memory management to compensate for the lack of garbage collection. The crossbeam crate is used for epoch-based memory management.

For details on the hash map's design and implementation, see the (private) [map_inner] module.

At the time of writing, other concurrent hash maps available don't appear to allow reading and writing at the same time. This map does. Effectively, this map has the same guarantees as having a certain amount of global variables that can be changed atomically.

Reexports

extern crate std as std;

extern crate crossbeam_epoch;

extern crate crossbeam_utils as crossbeam;

use std::prelude::v1::*;

use std::borrow::Borrow;

use std::collections::hash_map::RandomState;

use std::fmt;

use std::hash::Hash;

use atomic::AtomicBox;

use map_inner::KeyCompare;

use map_inner::MapInner;

use map_inner::Match;

use map_inner::PutValue;

use map_inner::ValueSlot;

Modules

atomic

map_inner
This module implements most of the logic behind the [::LockFreeHashMap].

Structs

Guard
A guard that keeps the current thread pinned.

LockFreeHashMap

Scope

Constants

COPY_CHUNK_SIZE

Functions

pin
Pins the current thread.

scope
Create a new scope, for deferred destructors.