Struct lockfreehashmap::atomic::AtomicPtr

pub struct AtomicPtr<T>(Atomic<T>);

Wrapper around [Atomic]. Represents a pointer that can be null at first but never again afterwards.

Methods

impl<T> AtomicPtr<T>

pub fn new(value: Option<T>) -> Self

pub fn load<'g>(&self, guard: &'g Guard) -> MaybeNull<'g, T>

pub fn relaxed_exists<'g>(&self, guard: &'g Guard) -> bool

Check if the pointer is not null (i.e. the value exists), using Ordering::Relaxed.

pub unsafe fn take<'g>(&self, guard: &'g Guard) -> Option<Box<T>>

Swaps out the current value, leaving null in its place. Equivalent to [Option::take()].

Unsafe

This is unsafe because there must be no other pointers that can ever access this AtomicPtr after this function is called. This is intended to be used for [Drop::drop()].

pub fn compare_and_set<'g>(
&self,
compare: MaybeNull<T>,
set: NotNull<'g, T>,
guard: &'g Guard
) -> Result<NotNull<'g, T>, (MaybeNull<'g, T>, NotNull<'g, T>)>

pub fn compare_null_and_set<'g>(
&self,
set: NotNull<'g, T>,
guard: &'g Guard
) -> Result<NotNull<'g, T>, (NotNull<'g, T>, NotNull<'g, T>)>

pub fn compare_null_and_set_owned<'g>(
&self,
set: NotNullOwned<T>,
guard: &'g Guard
) -> Result<NotNull<'g, T>, (NotNull<'g, T>, NotNullOwned<T>)>

pub fn compare_and_set_owned<'g>(
&self,
compare: MaybeNull<T>,
set: NotNullOwned<T>,
guard: &'g Guard
) -> Result<NotNull<'g, T>, (MaybeNull<'g, T>, NotNullOwned<T>)>

pub fn compare_and_set_owned_weak<'g>(
&self,
compare: MaybeNull<T>,
set: NotNullOwned<T>,
guard: &'g Guard
) -> Result<NotNull<'g, T>, (MaybeNull<'g, T>, NotNullOwned<T>)>

pub fn tag<'g>(&self, guard: &'g Guard)

pub fn is_tagged<'g>(&self, guard: &'g Guard) -> bool

pub unsafe fn try_drop(&mut self, guard: &Guard)